Fixes for new structure
