+++ alliander_robotics/predefined_configurations.py
@@ -361,3 +361,12 @@ class PredefinedConfigurations:
         vehicle = Vehicle("panther", (0, -0.5, 0.2))
         arm = Arm("franka", (0, 0.5, 0))
         self.plat_conf.platforms = [vehicle, arm]
+
+    # Openvla test
+    @register_configuration("openvla")
+    def config_openvla(self) -> None:  # noqa: D102
+        arm = Arm("franka", moveit=True, gripper=True, openvla=True)
+        camera = Camera("realsense", (0.05, 0, 0), (0, -90, 180))
+
+        link(arm, camera)
+        self.plat_conf.platforms = [arm, camera]
